修正dao层、service层的保存方法（book的主键isbn是个Long，方法返回值也应为Long）

+++ src/main/java/dao/impl/BookDaoImpl.java
@@ -10,8 +10,8 @@ import model.Book;
 @SuppressWarnings("deprecation")
 public class BookDaoImpl extends HibernateDaoSupport implements BookDao {
 
-	public Integer save(Book book) {
-		return (Integer) getHibernateTemplate().save(book);
+	public Long save(Book book) {
+		return (Long) getHibernateTemplate().save(book);
 	}
 
 	public void delete(Book book) {
